+++ be/src/exec/CMakeLists.txt
@@ -132,7 +132,7 @@ set(EXEC_FILES
     pipeline/exchange/local_exchange.cpp
     pipeline/exchange/local_exchange_sink_operator.cpp
     pipeline/exchange/local_exchange_source_operator.cpp
-    pipeline/exchange/mcast_local_exchange.cpp
+    pipeline/exchange/multi_cast_local_exchange.cpp
     pipeline/exchange/sink_buffer.cpp
     pipeline/fragment_executor.cpp
     pipeline/operator.cpp
